refactor(cloud): extract viewer config and auth hooks
